openvidu-server, openvidu-node-client, openvidu-java-client: Add COMPOSED_QUICK_START outputMode

+++ openvidu-server/src/main/java/io/openvidu/server/kurento/core/KurentoSessionManager.java
@@ -30,6 +30,7 @@ import java.util.concurrent.TimeUnit;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+import io.openvidu.java.client.*;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.kurento.client.GenericMediaElement;
 import org.kurento.client.IceCandidate;
@@ -47,11 +48,6 @@ import com.google.gson.JsonObject;
 import io.openvidu.client.OpenViduException;
 import io.openvidu.client.OpenViduException.Code;
 import io.openvidu.client.internal.ProtocolElements;
-import io.openvidu.java.client.MediaMode;
-import io.openvidu.java.client.RecordingLayout;
-import io.openvidu.java.client.RecordingMode;
-import io.openvidu.java.client.RecordingProperties;
-import io.openvidu.java.client.SessionProperties;
 import io.openvidu.server.core.EndReason;
 import io.openvidu.server.core.FinalUser;
 import io.openvidu.server.core.IdentifierPrefixes;
@@ -144,6 +140,12 @@ public class KurentoSessionManager extends SessionManager {
 				}
 			}
 
+			// If Recording default layout is COMPOSED_QUICK_START
+			Recording.OutputMode defaultOutputMode = kSession.getSessionProperties().defaultOutputMode();
+			if (defaultOutputMode.equals(Recording.OutputMode.COMPOSED_QUICK_START)) {
+				recordingManager.startComposedQuickStartContainer(kSession);
+			}
+
 			if (kSession.isClosed()) {
 				log.warn("'{}' is trying to join session '{}' but it is closing", participant.getParticipantPublicId(),
 						sessionId);
@@ -214,12 +216,12 @@ public class KurentoSessionManager extends SessionManager {
 						Participant p = sessionidParticipantpublicidParticipant.get(sessionId)
 								.remove(participant.getParticipantPublicId());
 
-						if (this.openviduConfig.isTurnadminAvailable()) {
+						if (p != null && this.openviduConfig.isTurnadminAvailable()) {
 							this.coturnCredentialsService.deleteUser(p.getToken().getTurnCredentials().getUsername());
 						}
 
 						// TODO: why is this necessary??
-						if (insecureUsers.containsKey(p.getParticipantPrivateId())) {
+						if (p != null && insecureUsers.containsKey(p.getParticipantPrivateId())) {
 							boolean stillParticipant = false;
 							for (Session s : sessions.values()) {
 								if (!s.isClosed()
@@ -295,6 +297,14 @@ public class KurentoSessionManager extends SessionManager {
 									"Last participant left. Starting {} seconds countdown for stopping recording of session {}",
 									this.openviduConfig.getOpenviduRecordingAutostopTimeout(), sessionId);
 							recordingManager.initAutomaticRecordingStopThread(session);
+
+						} else if (remainingParticipants.size() == 1 && openviduConfig.isRecordingModuleEnabled()
+								&& MediaMode.ROUTED.equals(session.getSessionProperties().mediaMode())
+								&& session.getSessionProperties().defaultOutputMode().equals(Recording.OutputMode.COMPOSED_QUICK_START)
+								&& ProtocolElements.RECORDER_PARTICIPANT_PUBLICID
+								.equals(remainingParticipants.iterator().next().getParticipantPublicId())) {
+							// If no recordings are active in COMPOSED_QUICK_START output mode, stop container
+							recordingManager.stopComposedQuickStartContainer(session, reason);
 						}
 					}
